% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_content_parsed}
\alias{get_content_parsed}
\title{Get content and parse it}
\usage{
get_content_parsed(url)
}
\arguments{
\item{url}{A string URL linking to the API. If it does not contain the base
"https://api.covid19tracker.ca", then \code{url} will be combined with the base
to attempt to make a valid URL (and return a warning).}
}
\value{
A list.
}
\description{
Sends a GET request to https://api.covid19tracker.ca/.
If the request is successful and the returned content is JSON, formats it and
returns it parsed (via \code{jsonlite::fromJSON}).
}
\examples{

get_content_parsed("https://api.covid19tracker.ca/provinces")

}
