% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provinces.R
\name{get_provinces}
\alias{get_provinces}
\title{Get provinces and territories}
\usage{
get_provinces()
}
\value{
A data frame with a row per province/territory.
}
\description{
Get provinces and territories
}
\examples{

get_provinces()

}
