% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{get_subregions}
\alias{get_subregions}
\title{Get sub-regions}
\usage{
get_subregions(subregion_code = NULL)
}
\arguments{
\item{subregion_code}{One or more sub-regions to be returned.}
}
\value{
A data frame with a row per sub-region.
}
\description{
Returns a list of sub-regions in the COVID-19 tracker.
By default, returns all 805 sub-regions.
}
\examples{

get_subregions()
get_subregions("AB001")
get_subregions(c("SK003", "SK005"))
}
