% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{get_summary}
\alias{get_summary}
\title{Get the most recent summary data}
\usage{
get_summary(split = c("overall", "province", "region"))
}
\arguments{
\item{split}{One of "overall", "province", or "region" to specify how the
data is split.}
}
\value{
A data frame containing the summary data.
}
\description{
Runs a GET request of summary data from the COVID-19 tracker API, and
returns parsed data.
Via the \code{split} argument, data my be "overall" (all provinces/territories
combined), by "province" (one row per province/territory) or by "region"
(one row per health region).
}
\examples{

get_summary()
get_summary("province")
get_summary("region")
}
